Support ClangFormat v7.0.0
Add an indenter configuration file for [ClangFormat v7.0.0](http://releases.llvm.org/7.0.0/tools/clang/docs/ClangFormat.html). It can generate all the documented `.clang-format` options, except the following:

 - `ExperimentalAutoDetectBinPacking`, because it is discouraged from use in config files.

 - `AlwaysBreakAfterDefinitionReturnType`, as it is now deprecated.

 - `IncludeCategories` and `RawStringFormats`, because their option values are highly structured YAML and beyond the capabilities of UniversalIndentGUI.

Also enlist ClangFormat in the list of supported indenters in the text, HTML and Markdown README files (but not in the binary PDF file) and in the release packaging scripts.